+++ packages/labs/src/common/classes.ts
@@ -5,12 +5,6 @@
  */
 
 export const INPUT_GHOST = "pt-input-ghost";
-export const MULTISELECT = "pt-multi-select";
-export const MULTISELECT_POPOVER = `${MULTISELECT}-popover`;
-export const OMNIBAR = "pt-omnibar";
-export const OMNIBAR_OVERLAY = `${OMNIBAR}-overlay`;
-export const SELECT = "pt-select";
-export const SELECT_POPOVER = `${SELECT}-popover`;
 export const TAG_INPUT = "pt-tag-input";
 export const TAG_INPUT_ICON = `${TAG_INPUT}-icon`;
 export const TIMEZONE_PICKER = "pt-timezone-picker";
