+++ packages/select/src/blueprint-select.scss
@@ -3,5 +3,4 @@
 // Licensed under the terms of the LICENSE file distributed with this project.
  */
 
-@import "tag-input/tag-input";
-@import "timezone-picker/timezone-picker";
+@import "./components/index";
