+++ packages/select/src/components/_index.scss
@@ -3,5 +3,6 @@
 // Licensed under the terms of the LICENSE file distributed with this project.
  */
 
-@import "tag-input/tag-input";
-@import "timezone-picker/timezone-picker";
+@import "omnibar/omnibar";
+@import "select/multi-select";
+@import "select/select";
