+++ packages/select/src/components/classes.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright 2017 Palantir Technologies, Inc. All rights reserved.
+ *
+ * Licensed under the terms of the LICENSE file distributed with this project.
+ */
+
+export const MULTISELECT = "pt-multi-select";
+export const MULTISELECT_POPOVER = `${MULTISELECT}-popover`;
+export const OMNIBAR = "pt-omnibar";
+export const OMNIBAR_OVERLAY = `${OMNIBAR}-overlay`;
+export const SELECT = "pt-select";
+export const SELECT_POPOVER = `${SELECT}-popover`;
