+++ packages/select/src/components/index.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright 2017 Palantir Technologies, Inc. All rights reserved.
+ *
+ * Licensed under the terms of the LICENSE file distributed with this project.
+ */
+
+export * from "./classes";
+export * from "./omnibar/omnibar";
+export * from "./query-list/queryList";
+export * from "./select/multiSelect";
+export * from "./select/select";
+export * from "./select/suggest";
