+++ packages/select/src/components/omnibar/_omnibar.scss
