+++ packages/select/src/components/omnibar/omnibar.tsx
@@ -21,7 +21,7 @@ import {
     Utils,
 } from "@blueprintjs/core";
 
-import * as Classes from "../../common/classes";
+import * as Classes from "../classes";
 import { IListItemsProps, IQueryListRendererProps, QueryList } from "../query-list/queryList";
 import { ISelectItemRendererProps } from "../select/select";
