+++ packages/select/src/components/select/_multi-select.scss
