+++ packages/select/src/components/select/_select.scss
