+++ packages/select/src/components/select/multiSelect.tsx
@@ -8,7 +8,7 @@ import * as PureRender from "pure-render-decorator";
 import * as React from "react";
 
 import { HTMLInputProps, IPopoverProps, Keys, Menu, Popover, Position, Utils } from "@blueprintjs/core";
-import * as Classes from "../../common/classes";
+import * as Classes from "../classes";
 import { IListItemsProps, IQueryListRendererProps, QueryList } from "../query-list/queryList";
 import { ISelectItemRendererProps } from "../select/select";
 import { ITagInputProps, TagInput } from "../tag-input/tagInput";
