+++ packages/select/src/index.md
@@ -1,0 +1,23 @@
+---
+reference: select
+---
+
+@# Select
+
+The **[@blueprintjs/select](https://www.npmjs.com/package/@blueprintjs/select)** NPM package provides React components related to selecting items from a list:
+
+- [`Select`](#labs/select-component) for selecting items in a list.
+
+- [`Suggest`](#labs/suggest) for selecting items in a list, from a text input.
+
+- [`MultiSelect`](#labs/multi-select) for selecting multiple items in a list.
+
+- [`Omnibar`](#labs/omnibar), a macOS spotlight-style typeahead component.
+
+- [`QueryList`](#labs/query-list), a higher-order component that provides interactions between a query string and a list of items.
+
+@page select-component
+@page suggest
+@page multi-select
+@page omnibar
+@page query-list
