+++ packages/select/src/index.ts
@@ -1,0 +1,7 @@
+/*
+ * Copyright 2017 Palantir Technologies, Inc. All rights reserved.
+ *
+ * Licensed under the terms of the LICENSE file distributed with this project.
+ */
+
+export * from "./components";
