+++ packages/select/src/tsconfig.cjs.json
@@ -1,0 +1,7 @@
+{
+    "extends": "./tsconfig",
+    "compilerOptions": {
+        "module": "commonjs",
+        "outDir": "../dist"
+    }
+}
