+++ packages/select/src/tsconfig.json
@@ -1,0 +1,6 @@
+{
+    "extends": "../../../config/tsconfig.base",
+    "compilerOptions": {
+        "outDir": "../dist/esm"
+    }
+}
